+++ RadioApp/Modules/Onboarding Screen/OnboardingPage1VC.swift
@@ -1,17 +1,18 @@
 //
-//  OnboardingPage1VC.swift
+//  OnboardingVC.swift
 //  RadioApp
 //
-//  Created by Andrew Linkov on 03.08.2024.
+//  Created by Evgenii Mazrukho on 29.07.2024.
 //
 
 import UIKit
 import SnapKit
 
-final class OnboardingPage1VC: UIPageViewController {
+final class OnboardingPage1VC: UIViewController {
     
     var pages = [UIViewController]()
-
+    let pageControl = UIPageControl()
+    let initialPage = 0
     
     // MARK: - UI Properties
     private lazy var backgroundImageView: UIImageView = {
@@ -21,82 +22,51 @@ final class OnboardingPage1VC: UIPageViewController {
         return element
     }()
     
-    private var gradientLayer: CAGradientLayer {
-        let element = CAGradientLayer()
-        element.frame = view.bounds
-        element.colors = [UIColor.black.withAlphaComponent(0.9).cgColor, UIColor.clear.cgColor]
-        element.startPoint = CGPoint(x: 0.5, y: 1.0)
-        element.endPoint = CGPoint(x: 0.5, y: 0.0)
-        return element
-    }
-    
-    private lazy var skipButton: UIButton = {
-        let element = UIButton()
-        element.setTitle("Skip", for: .normal)
-        element.titleLabel?.font = UIFont.systemFont(ofSize: 18, weight: .regular)
-        element.addTarget(self, action: #selector(skipButtonTapped), for: .touchUpInside)
-        return element
-    }()
-    
-    private lazy var continueButton: UIButton = {
-        let element = UIButton()
-        element.setTitle("Continue", for: .normal)
-        element.titleLabel?.font = UIFont.systemFont(ofSize: 24, weight: .medium)
-        element.backgroundColor = UIColor(named: "buttonColor")
-        element.layer.cornerRadius = 25
-        element.addTarget(self, action: #selector(continueButtonTapped), for: .touchUpInside)
-        return element
-    }()
-    
-    private lazy var progressStackView: UIStackView = {
+    private lazy var headerStackView: UIStackView = {
         let element = UIStackView()
         element.axis = .horizontal
-        element.spacing = 10
+        element.alignment = .center
+        element.spacing = 5
         element.distribution = .fillProportionally
         return element
     }()
     
-    private lazy var firstProgressView: UIView = {
-        let element = UIView()
-        element.backgroundColor = UIColor(red: 0.98, green: 0.61, blue: 0.69, alpha: 1.00)
-        element.layer.cornerRadius = 5
-        return element
-    }()
-    
-    private lazy var secondProgressView: UIView = {
-        let element = UIView()
-        element.backgroundColor = .lightGray
-        element.layer.cornerRadius = 5
-        return element
-    }()
-    
-    private lazy var thirdProgressView: UIView = {
-        let element = UIView()
-        element.backgroundColor = .lightGray
-        element.layer.cornerRadius = 5
-        return element
-    }()
-    
-    private lazy var mainLabel: UILabel = {
+    private lazy var firstHeaderLabel: UILabel = {
         let element = UILabel()
-        element.textAlignment = .center
-        element.font = .systemFont(ofSize: 55, weight: .bold)
+        element.text = "Let's Get Started"
+        element.font = .systemFont(ofSize: 60, weight: .bold)
         element.textColor = .white
-        element.numberOfLines = 0
-        let fullString = "LET'S GET STARTED"
         return element
     }()
     
-    weak var coordinator: OnboardingCoordinator!
+    private lazy var secondHeaderLabel: UILabel = {
+        let element = UILabel()
+        element.text = "Enjoy the best radio stations from your home? don't miss out on anything"
+        element.font = .systemFont(ofSize: 16, weight: .regular)
+        element.textColor = .white
+        return element
+    }()
     
-    init(coordinator: OnboardingCoordinator!) {
-        super.init(transitionStyle: .scroll, navigationOrientation: .horizontal, options: nil)
-        self.coordinator = coordinator
-    }
-    
-    required init?(coder: NSCoder) {
-        fatalError("init(coder:) has not been implemented")
-    }
+    private lazy var getStartedButton: UIButton = {
+        let element = UIButton()
+        element.setTitle("Get started", for: .normal)
+        element.titleLabel?.font = UIFont.systemFont(ofSize: 20, weight: .medium)
+        element.backgroundColor = UIColor(named: "pinkBase")
+        element.layer.cornerRadius = 10
+        element.addTarget(self, action: #selector(continueButtonTapped), for: .touchUpInside)
+        return element
+    }()
+
+    //    weak var coordinator: OnboardingCoordinator!
+    //
+    //    init(coordinator: OnboardingCoordinator!) {
+    //        super.init(nibName: nil, bundle: nil)
+    //        self.coordinator = coordinator
+    //    }
+    //
+    //    required init?(coder: NSCoder) {
+    //        fatalError("init(coder:) has not been implemented")
+    //    }
     
     // MARK: - Life Cycle
     override func viewDidLoad() {
@@ -108,21 +78,16 @@ final class OnboardingPage1VC: UIPageViewController {
     // MARK: - Private Methods
     private func setupUI() {
         view.addSubview(backgroundImageView)
-        view.layer.addSublayer(gradientLayer)
-        view.addSubview(skipButton)
-        view.addSubview(continueButton)
-        view.addSubview(progressStackView)
-        view.addSubview(mainLabel)
+        view.addSubview(headerStackView)
+        view.addSubview(getStartedButton)
         
-        
-        progressStackView.addArrangedSubview(firstProgressView)
-        progressStackView.addArrangedSubview(secondProgressView)
-        progressStackView.addArrangedSubview(thirdProgressView)
+        headerStackView.addArrangedSubview(firstHeaderLabel)
+        headerStackView.addArrangedSubview(secondHeaderLabel)
     }
     // MARK: - Selector methods
     @objc private func continueButtonTapped() {
         print("continueButtonTapped")
-        navigationController?.pushViewController(OnboardingPage2VC (coder: coordinator) ?? <#default value#>, animated: true)
+        navigationController?.pushViewController(OnboardingPage2VC(), animated: true)
     }
     
     @objc private func skipButtonTapped() {
@@ -132,47 +97,24 @@ final class OnboardingPage1VC: UIPageViewController {
 
 // MARK: - Setup Constraints
 extension OnboardingPage1VC {
-    private func setupConstraints() {
+    func setupConstraints() {
         backgroundImageView.snp.makeConstraints { make in
             make.edges.equalToSuperview()
         }
         
-        skipButton.snp.makeConstraints { make in
-            make.bottom.equalTo(view.safeAreaLayoutGuide).inset(20)
+        headerStackView.snp.makeConstraints { make in
             make.centerX.equalToSuperview()
+            make.top.equalTo(view.safeAreaLayoutGuide).inset(20)
         }
         
-        continueButton.snp.makeConstraints { make in
+        getStartedButton.snp.makeConstraints { make in
+            make.bottom.equalTo(view.safeAreaLayoutGuide).inset(40)
+            make.height.equalTo(80)
+            make.width.equalTo(310)
             make.centerX.equalToSuperview()
-            make.bottom.equalTo(skipButton).inset(50)
-            make.height.equalTo(50)
-            make.width.equalTo(view.frame.width / 1.8)
-        }
-        
-        progressStackView.snp.makeConstraints { make in
-            make.centerX.equalToSuperview()
-            make.bottom.equalTo(continueButton).inset(90)
-        }
-        
-        firstProgressView.snp.makeConstraints { make in
-            make.height.equalTo(10)
-            make.width.equalTo(50)
-        }
-        
-        secondProgressView.snp.makeConstraints { make in
-            make.height.equalTo(10)
-            make.width.equalTo(50)
-        }
-        
-        thirdProgressView.snp.makeConstraints { make in
-            make.height.equalTo(10)
-            make.width.equalTo(50)
-        }
-        
-        mainLabel.snp.makeConstraints { make in
-            make.centerX.equalToSuperview()
-            make.bottom.equalTo(progressStackView).inset(40)
         }
     }
 }
+
+
