--- RadioApp/Modules/Onboarding Screen/OnboardingPageVC.swift
@@ -1,70 +1,0 @@
-//
-//  OnboardingPageVC.swift
-//  RadioApp
-//
-//  Created by Andrew Linkov on 03.08.2024.
-//
-
-import UIKit
-
-final class OnboardingPageViewController: UIPageViewController, UIPageViewControllerDataSource, UIPageViewControllerDelegate {
-    
-    var pageControllers: [UIViewController] = []
-    weak var coordinator: OnboardingCoordinator!
-    
-    init(coordinator: OnboardingCoordinator!) {
-        super.init(transitionStyle: .scroll, navigationOrientation: .horizontal, options: nil)
-        self.coordinator = coordinator
-    }
-    
-    required init?(coder: NSCoder) {
-        fatalError("init(coder:) has not been implemented")
-    }
-    
-    override func viewDidLoad() {
-        
-        super.viewDidLoad()
-        
-        dataSource = self
-        delegate  = self
-        
-        let firstController = OnboardingMainVC (coder: coordinator)
-        let secondController = OnboardingPage1VC (coder: coordinator)
-        let thirdController = OnboardingPage2VC (coder: coordinator)
-        
-        pageControllers = [firstController, secondController, thirdController]
-        
-        if let firstVC = pageControllers.first {
-            setViewControllers([firstVC], direction: .forward, animated: true, completion: nil)
-        }
-    }
-    
-    func pageViewController(_ pageViewController: UIPageViewController, viewControllerBefore viewController: UIViewController) -> UIViewController? {
-        guard let currentIndex = pageControllers.firstIndex(of: viewController) else {
-            return nil
-        }
-        
-        let previousIndex = currentIndex - 1
-        
-        guard previousIndex >= 0 else {
-            return nil
-        }
-        
-        return pageControllers[previousIndex]
-    }
-    
-    func pageViewController(_ pageViewController: UIPageViewController, viewControllerAfter viewController: UIViewController) -> UIViewController? {
-        guard let currentIndex = pageControllers.firstIndex(of: viewController) else {
-            return nil
-        }
-        
-        let nextIndex = currentIndex + 1
-        
-        guard nextIndex < pageControllers.count else {
-            return nil
-        }
-        
-        return pageControllers[nextIndex]
-    }
-}
-
